+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,12 +13,31 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('laravel_users', function (Blueprint $table) {
+        Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
+            $table->string('username')->unique();
+            $table->string('nom');
+            $table->string('prenom');
+            $table->string('email');
+            $table->string('pays');
+            $table->string('numero');
             $table->string('password');
+            $table->boolean('active')->default(false);
+            $table->timestamp('date_inscription')->nullable();
+            $table->string('ville')->nullable();
+            $table->string('commune')->nullable();
+            $table->string('promo1')->nullable();
+            $table->string('promo2')->nullable();
+            $table->string('emploi')->nullable();
+            $table->string('universite')->nullable();
+            $table->boolean('is_staff')->default(false);
+            $table->string('staff_role')->default("member");
+            $table->unsignedInteger('version')->default(3);
+            $table->string('photo')->nullable(); 
+
+            $table->boolean('is_superadmin')->default(false); // Version 6
+
+            $table->timestamp('email_verified_at')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });